Refactor author components; update return types in documentation, streamline imports, and enhance styling
