+++ src/js/components/index.js
@@ -1,0 +1,7 @@
+/**
+ * Re-export all components for simpler imports
+ */
+export { default as AuthorSelector } from './AuthorSelector';
+export { default as AuthorPicker } from './AuthorPicker';
+export { default as AuthorItem } from './AuthorItem';
+export { default as AuthorsListPreview } from './AuthorsListPreview';
