+++ src/js/hooks/index.js
@@ -1,0 +1,5 @@
+/**
+ * Re-export all hooks for simpler imports
+ */
+export { default as useAuthors } from './useAuthors';
+export { default as useAuthorsList } from './useAuthorsList';
