+++ src/js/index.js
@@ -1,0 +1,10 @@
+/**
+ * Main entry point for js utilities
+ * Export everything from different modules for easier imports
+ */
+
+// Re-export modules
+export * from './components';
+export * from './hooks';
+export * from './services';
+export * from './utils';
