+++ src/js/services/index.js
@@ -1,0 +1,4 @@
+/**
+ * Re-export all services for simpler imports
+ */
+export * from './api';
