+++ src/js/utils/index.js
@@ -1,0 +1,5 @@
+/**
+ * Re-export utility functions for simpler imports
+ */
+export * from './author-utils';
+export * from './block-utils';
